Removed updated image-processing.py and removed updated-image-processing.py
